+++ docs/changelog.md
@@ -1,5 +1,18 @@
 # CoconutGo Changelog
 
+## 0.12.4
+
+* Added option to specify Service Providers in client's config
+
+## 0.12.3
+
+* Added timeout to tendermint monitor startup. It is required for fresh chains that have not gone past genesis block
+* Button to explicitly re-randomize credential
+* Updating balance on load
+* Other minor changes and fixes
+* Configuration for nymnet hosted on AWS + scripts for automatically copying data over
+* Makefile additions
+
 ## 0.12.2
 
 * Fixed Dockerfile for the issuer